refactoring: common buffer.js, readPBFElementFromBuffer now the same; could also be inlined, but used multiple times, preferring to keep central function?
