+++ lib/browser/arrayBufferReader.js
@@ -1,4 +1,4 @@
-var buf = require('./buffer.js');
+var buf = require('../buffer.js');
 require("setimmediate");
 
 function readBlobHeaderSize(fd, position, size, callback){
