+++ lib/browser/fileReader.js
@@ -1,4 +1,4 @@
-var buf = require('./buffer.js');
+var buf = require('../buffer.js');
 
 function readBlobHeaderSize(file, position, size, callback){
     read(file, position, size, function(err, buffer){
