+++ lib/buffer.js
