--- lib/nodejs/buffer.js
@@ -1,7 +1,0 @@
-function readPBFElementFromBuffer(buffer, pbfDecode, callback){
-    return callback(null, pbfDecode(buffer));
-}
-
-module.exports = {
-    readPBFElementFromBuffer: readPBFElementFromBuffer
-};
