+++ lib/nodejs/fsReader.js
@@ -1,5 +1,5 @@
 var fs = require('fs');
-var buf = require('./buffer.js');
+var buf = require('../buffer.js');
 
 function bytesReadFail(callback, expectedBytes, readBytes){
     return callback(new Error('Expected ' + expectedBytes + ' bytes but got ' + readBytes));
