+++ lib/pbfParser.js
@@ -11,18 +11,17 @@
 var blockFormat = require('./proto/osmformat.js');
 var fileFormat = require('./proto/fileformat.js');
 var protoBuf = require("protobufjs");
+var buf = require('./buffer.js');
 
-var zlib, buf, reader, arrayBufferReader, fileReader;
+var zlib, reader, arrayBufferReader, fileReader;
 
 // check if running in Browser or Node.js (use self not window because of Web Workers)
 if (typeof self !== 'undefined') {
     zlib = require('./browser/zlib.js');
-    buf = require('./browser/buffer.js');
     arrayBufferReader = require('./browser/arrayBufferReader.js');
     fileReader = require('./browser/fileReader.js');
 } else {
     zlib = require('./nodejs/zlib.js');
-    buf = require('./nodejs/buffer.js');
     reader = require('./nodejs/fsReader.js');
 }
